# Performance optimization: using NextBytesNoCopy where possible