--- a/data_psi_test.go
+++ b/data_psi_test.go
@@ -374,3 +374,10 @@ func TestWritePSIData(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkParsePSIData(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		parsePSIData(astikit.NewBytesIterator(psiBytes()))
+	}
+}