--- a/go.mod
+++ b/go.mod
@@ -7,3 +7,5 @@ require (
 	github.com/pkg/profile v1.4.0
 	github.com/stretchr/testify v1.4.0
 )
+
+replace github.com/asticode/go-astikit v0.19.0 => github.com/barbashov/go-astikit v0.19.1