(For this to merge, https://github.com/asticode/go-astikit/pull/9 should be merged before.)

I've changed a bunch of places in demuxer to use NextByteNoCopy instead of NextByte.
I can't say it's night and day difference, but still it's a nice little improvement :)
Benchmark results follow.

Before:
```
goos: darwin
goarch: arm64
pkg: github.com/asticode/go-astits
BenchmarkParsePATSection-8   	12254799	        97.23 ns/op	      72 B/op	       7 allocs/op
BenchmarkParsePESData
BenchmarkParsePESData/without_header-8         	17240697	        69.27 ns/op	      56 B/op	       4 allocs/op
BenchmarkParsePESData/with_header-8            	 4049978	       295.6 ns/op	     304 B/op	      18 allocs/op
BenchmarkParsePMTSection-8                     	 4220971	       282.7 ns/op	     632 B/op	      15 allocs/op
BenchmarkParsePSIData-8                        	  141775	      8532 ns/op	    5321 B/op	     169 allocs/op
BenchmarkParseDescriptor
BenchmarkParseDescriptor/AC3-8                                	 8806401	       136.0 ns/op	     288 B/op	       6 allocs/op
BenchmarkParseDescriptor/ISO639LanguageAndAudioType-8         	 9917152	       121.1 ns/op	     288 B/op	       6 allocs/op
BenchmarkParseDescriptor/MaximumBitrate-8                     	10645449	       113.2 ns/op	     260 B/op	       6 allocs/op
BenchmarkParseDescriptor/NetworkName-8                        	 9511978	       126.6 ns/op	     280 B/op	       6 allocs/op
BenchmarkParseDescriptor/Service-8                            	 8413123	       141.9 ns/op	     336 B/op	       7 allocs/op
BenchmarkParseDescriptor/ShortEvent-8                         	 7699062	       156.6 ns/op	     344 B/op	       8 allocs/op
BenchmarkParseDescriptor/StreamIdentifier-8                   	11814889	       101.9 ns/op	     254 B/op	       5 allocs/op
BenchmarkParseDescriptor/Subtitling-8                         	 4409683	       271.7 ns/op	     412 B/op	      15 allocs/op
BenchmarkParseDescriptor/Teletext-8                           	 5119393	       234.3 ns/op	     370 B/op	      11 allocs/op
BenchmarkParseDescriptor/ExtendedEvent-8                      	 5383294	       222.2 ns/op	     416 B/op	      11 allocs/op
BenchmarkParseDescriptor/EnhancedAC3-8                        	 8112375	       147.8 ns/op	     304 B/op	       6 allocs/op
BenchmarkParseDescriptor/Extension-8                          	 6975066	       171.7 ns/op	     352 B/op	       8 allocs/op
BenchmarkParseDescriptor/Component-8                          	 8063560	       148.5 ns/op	     324 B/op	       7 allocs/op
BenchmarkParseDescriptor/Content-8                            	 7474143	       161.0 ns/op	     290 B/op	       8 allocs/op
BenchmarkParseDescriptor/ParentalRating-8                     	 7275045	       164.7 ns/op	     320 B/op	       8 allocs/op
BenchmarkParseDescriptor/LocalTimeOffset-8                    	 2556608	       468.0 ns/op	     392 B/op	      13 allocs/op
BenchmarkParseDescriptor/VBIData-8                            	 6399276	       187.6 ns/op	     326 B/op	       9 allocs/op
BenchmarkParseDescriptor/VBITeletext-8                        	 7012162	       170.5 ns/op	     320 B/op	       8 allocs/op
BenchmarkParseDescriptor/AVCVideo-8                           	10797754	       111.3 ns/op	     264 B/op	       5 allocs/op
BenchmarkParseDescriptor/PrivateDataSpecifier-8               	10777999	       112.2 ns/op	     260 B/op	       6 allocs/op
BenchmarkParseDescriptor/DataStreamAlignment-8                	11817196	       101.8 ns/op	     254 B/op	       5 allocs/op
BenchmarkParseDescriptor/PrivateDataIndicator-8               	10707218	       112.0 ns/op	     260 B/op	       6 allocs/op
BenchmarkParseDescriptor/UserDefined-8                        	12038341	       100.3 ns/op	     256 B/op	       5 allocs/op
BenchmarkParseDescriptor/Registration-8                       	 8706368	       137.6 ns/op	     292 B/op	       7 allocs/op
BenchmarkParseDescriptor/Unknown-8                            	 9858508	       121.5 ns/op	     288 B/op	       6 allocs/op
BenchmarkParseDescriptor/Extension#01-8                       	 8554720	       140.0 ns/op	     304 B/op	       7 allocs/op
BenchmarkParsePacket-8                                        	 4242537	       281.3 ns/op	     448 B/op	      15 allocs/op
BenchmarkDemuxer_NextData-8   	     	     	     	     	  179368	        6088 ns/op	    6410 B/op	     188 allocs/op
```

After:
```
goos: darwin
goarch: arm64
pkg: github.com/asticode/go-astits
BenchmarkParsePATSection-8      14293448            83.59 ns/op       64 B/op          5 allocs/op
BenchmarkParsePESData
BenchmarkParsePESData/without_header-8          20300631            58.82 ns/op       52 B/op          3 allocs/op
BenchmarkParsePESData/with_header-8              5215488           229.1 ns/op       280 B/op         10 allocs/op
BenchmarkParsePMTSection-8                       5153559           232.2 ns/op       618 B/op          9 allocs/op
BenchmarkParsePSIData-8                           150284          7986 ns/op        5049 B/op        115 allocs/op
BenchmarkParseDescriptor
BenchmarkParseDescriptor/AC3-8                                   9969420           119.7 ns/op       284 B/op          4 allocs/op
BenchmarkParseDescriptor/ISO639LanguageAndAudioType-8           11080315           108.2 ns/op       284 B/op          4 allocs/op
BenchmarkParseDescriptor/MaximumBitrate-8                       14564409            82.07 ns/op      252 B/op          3 allocs/op
BenchmarkParseDescriptor/NetworkName-8                          11150904           108.1 ns/op       276 B/op          4 allocs/op
BenchmarkParseDescriptor/Service-8                               9185078           130.7 ns/op       328 B/op          5 allocs/op
BenchmarkParseDescriptor/ShortEvent-8                            8483942           141.8 ns/op       344 B/op          6 allocs/op
BenchmarkParseDescriptor/StreamIdentifier-8                     15024884            79.83 ns/op      249 B/op          3 allocs/op
BenchmarkParseDescriptor/Subtitling-8                            5325788           226.1 ns/op       398 B/op          9 allocs/op
BenchmarkParseDescriptor/Teletext-8                              5553873           215.2 ns/op       366 B/op          9 allocs/op
BenchmarkParseDescriptor/ExtendedEvent-8                         5627692           209.0 ns/op       416 B/op          9 allocs/op
BenchmarkParseDescriptor/EnhancedAC3-8                           9031544           132.8 ns/op       300 B/op          4 allocs/op
BenchmarkParseDescriptor/Extension-8                             7636591           157.5 ns/op       352 B/op          6 allocs/op
BenchmarkParseDescriptor/Component-8                             9726552           123.7 ns/op       320 B/op          5 allocs/op
BenchmarkParseDescriptor/Content-8                               9240913           129.0 ns/op       283 B/op          5 allocs/op
BenchmarkParseDescriptor/ParentalRating-8                        8024780           150.0 ns/op       316 B/op          6 allocs/op
BenchmarkParseDescriptor/LocalTimeOffset-8                       2829170           424.7 ns/op       376 B/op          7 allocs/op
BenchmarkParseDescriptor/VBIData-8                               6965751           172.3 ns/op       322 B/op          7 allocs/op
BenchmarkParseDescriptor/VBITeletext-8                           7676050           156.5 ns/op       315 B/op          6 allocs/op
BenchmarkParseDescriptor/AVCVideo-8                             13757174            87.12 ns/op      256 B/op          3 allocs/op
BenchmarkParseDescriptor/PrivateDataSpecifier-8                 14734185            82.31 ns/op      252 B/op          3 allocs/op
BenchmarkParseDescriptor/DataStreamAlignment-8                  14950507            80.34 ns/op      249 B/op          3 allocs/op
BenchmarkParseDescriptor/PrivateDataIndicator-8                 14538712            82.22 ns/op      252 B/op          3 allocs/op
BenchmarkParseDescriptor/UserDefined-8                          14633461            82.71 ns/op      252 B/op          3 allocs/op
BenchmarkParseDescriptor/Registration-8                         11132430           108.7 ns/op       284 B/op          4 allocs/op
BenchmarkParseDescriptor/Unknown-8                              11541582           104.9 ns/op       284 B/op          4 allocs/op
BenchmarkParseDescriptor/Extension#01-8                          9737190           123.8 ns/op       300 B/op          5 allocs/op
BenchmarkParsePacket-8                                           5243634           229.4 ns/op       416 B/op          9 allocs/op
BenchmarkDemuxer_NextData-8   	     	     	     	     	  206743	        5514 ns/op	    6082 B/op	     122 allocs/op
```