@asticode 2021-03-27
Now that `v0.20.0` is available for `astikit`, mod files need to be updated